+++ IDisposableAnalyzers/Helpers/Disposable.Stores.cs
@@ -51,6 +51,10 @@
         {
             switch (candidate)
             {
+                case { Parent: MemberAccessExpressionSyntax { Parent: InvocationExpressionSyntax { ArgumentList: { Arguments: { Count: 1 } arguments } } invocation } }
+                    when invocation.IsSymbol(KnownSymbol.DisposableMixins.DisposeWith, recursion.SemanticModel, recursion.CancellationToken) &&
+                         recursion.SemanticModel.TryGetSymbol(arguments[0].Expression, recursion.CancellationToken, out container):
+                    return true;
                 case { Parent: InitializerExpressionSyntax { Parent: ImplicitArrayCreationExpressionSyntax arrayCreation } }:
                     return StoresOrAssigns(arrayCreation, out container);
                 case { Parent: InitializerExpressionSyntax { Parent: ArrayCreationExpressionSyntax arrayCreation } }: